requirements.txt
catboost==1.1.1
contourpy==1.0.7
cycler==0.11.0
flake8==6.0.0
fonttools==4.38.0
graphviz==0.20.1
h5py==3.8.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.7.0
mccabe==0.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
tenacity==8.2.1
threadpoolctl==3.1.0
zipp==3.15.0
